Develop sibs

+++ app/Http/Controllers/WarehousesController.php
@@ -126,9 +126,12 @@ class WarehousesController extends Controller
 
     public function destroy(Warehouses $warehouse)
     {
-        $delete = $warehouse->delete();
+        if($warehouse){
+             $warehouse->products()->delete();
+             $delete = $warehouse->delete();
+        }
         if($delete){
-            return redirect()->route('dashboard.warehouses.index')->withSuccessMessage('Supplier Deleted Successfully');
+            return redirect()->route('dashboard.warehouses.index')->withSuccessMessage('Warehouse Deleted Successfully');
         }else{
             return redirect()->route('dashboard.warehouses.index')->withErrorMessage('AN Error Occured');
         }
